Move TORCH itemacro to its own file.
